Make result of `notification.cost()` into a float
Cost was returning a `Decimal`, which jsonify doesn't like.
Making it a float fixes the problem.

[Relevant issue on github](https://github.com/pallets/flask/issues/835).

+++ app/models.py
@@ -557,7 +557,7 @@ class Notification(db.Model):
             desc(ProviderRates.valid_from)
         ).limit(1).one()
 
-        return provider_rate.rate * self.billable_units
+        return float(provider_rate.rate * self.billable_units)
 
     def completed_at(self):
         if self.status in [
